Move Transaction in Progress (#737)
* move files

* fix tests
