+++ quadratic-core/src/controller.rs
@@ -1,12 +1,9 @@
 #[cfg(feature = "js")]
 use wasm_bindgen::prelude::*;
 
-use crate::grid::Grid;
+use crate::{computation::TransactionInProgress, grid::Grid};
 
-use self::{
-    dependencies::Dependencies, transaction_in_progress::TransactionInProgress,
-    transactions::Transaction,
-};
+use self::{dependencies::Dependencies, transactions::Transaction};
 
 pub mod auto_complete;
 pub mod borders;
@@ -21,7 +18,6 @@ pub mod operation;
 pub mod operations;
 pub mod sheet_offsets;
 pub mod sheets;
-pub mod transaction_in_progress;
 pub mod transaction_summary;
 pub mod transaction_types;
 pub mod transactions;
@@ -53,5 +49,15 @@ impl GridController {
     }
     pub fn grid(&self) -> &Grid {
         &self.grid
+    }
+    pub fn grid_mut(&mut self) -> &mut Grid {
+        &mut self.grid
+    }
+}
+
+#[cfg(test)]
+impl GridController {
+    pub fn get_transaction_in_progress(&self) -> Option<&TransactionInProgress> {
+        self.transaction_in_progress.as_ref()
     }
 }
